Cleanup.

+++ cuda_src/ALS_CUDA.cu
@@ -395,14 +395,15 @@ cudaError_t als_NV(smat_t& R_C, testset_t& T, mat_t& W, mat_t& H, parameter& par
         double f_rmse = 0;
         gpuErrchk(cudaMemcpy(rmse, d_rmse, (T.nnz + 1) * sizeof(float), cudaMemcpyDeviceToHost));
 
-        for (unsigned i = 0; i < T.nnz; ++i) {
+        for (long i = 0; i < T.nnz; ++i) {
             tot_rmse += rmse[i];
         }
         f_rmse = sqrtf(tot_rmse / T.nnz);
         rmse_timer.Stop();
 
         float rmse_time = rmse_timer.Elapsed();
-        printf("[-INFO-] iteration num %d \tupdate_time %.4lf|%.4lfs \tRMSE=%lf time:%fs\n", iter, update_time, update_time_acc, f_rmse, rmse_time);
+        printf("[-INFO-] iteration num %d \tupdate_time %.4lf|%.4lfs \tRMSE=%lf time:%fs\n",
+               iter, update_time, update_time_acc, f_rmse, rmse_time);
     }
 
     cudaStatus = cudaMemcpy(H_, dev_H_, nbits_H_, cudaMemcpyDeviceToHost);
